feat(backend, app, docs): 루틴 공유 기능 1차

+++ backend/src/main/java/com/back/domain/stats/dto/HomeStatsResponse.java
@@ -50,7 +50,7 @@ public record HomeStatsResponse(
     }
 
     public record SetDetail(
-            int weight,
+            double weight,
             int reps
     ) {
     }
