+++ backend/src/main/java/com/back/domain/workout/session/dto/WorkoutSetDetailResponse.java
@@ -5,7 +5,7 @@ import com.back.domain.workout.set.entity.WorkoutSet;
 public record WorkoutSetDetailResponse(
         Long setId,
         int setNumber,
-        int weight,
+        double weight,
         int reps
 ) {
     public static WorkoutSetDetailResponse from(WorkoutSet ws) {
